+++ examples/service/verify/messaging_configuration/delete/messaging_configuration_delete_example.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/RJPearson94/twilio-sdk-go"
 	v2 "github.com/RJPearson94/twilio-sdk-go/service/verify/v2"
-	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/webhooks"
 	"github.com/RJPearson94/twilio-sdk-go/session/credentials"
 )
 
@@ -25,18 +24,14 @@ func init() {
 }
 
 func main() {
-	resp, err := verifyClient.
+	err := verifyClient.
 		Service("VAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX").
-		Webhooks.
-		Create(&webhooks.CreateWebhookInput{
-			FriendlyName: "Test Rate Limit",
-			EventTypes:   []string{"*"},
-			WebhookURL:   "https://example.com/webhook",
-		})
+		MessagingConfiguration("all").
+		Delete()
 
 	if err != nil {
 		log.Panicf("%s", err.Error())
 	}
 
-	log.Printf("SID: %s", resp.Sid)
+	log.Println("Messaging configuration deleted")
 }