+++ examples/service/verify/messaging_configuration/pagination/messaging_configuration_pagination_example.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/RJPearson94/twilio-sdk-go"
 	v2 "github.com/RJPearson94/twilio-sdk-go/service/verify/v2"
-	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/webhooks"
 	"github.com/RJPearson94/twilio-sdk-go/session/credentials"
 )
 
@@ -25,18 +24,19 @@ func init() {
 }
 
 func main() {
-	resp, err := verifyClient.
+	paginator := verifyClient.
 		Service("VAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX").
-		Webhooks.
-		Create(&webhooks.CreateWebhookInput{
-			FriendlyName: "Test Rate Limit",
-			EventTypes:   []string{"*"},
-			WebhookURL:   "https://example.com/webhook",
-		})
+		MessagingConfigurations.
+		NewMessagingConfigurationsPaginator()
 
-	if err != nil {
-		log.Panicf("%s", err.Error())
+	for paginator.Next() {
+		currentPage := paginator.CurrentPage()
+		log.Printf("%v messaging configuration(s) found on page %v", len(currentPage.MessagingConfigurations), currentPage.Meta.Page)
 	}
 
-	log.Printf("SID: %s", resp.Sid)
+	if paginator.Error() != nil {
+		log.Panicf("%s", paginator.Error())
+	}
+
+	log.Printf("Total number of messaging configuration(s) found: %v", len(paginator.MessagingConfigurations))
 }